Cleanup imports, console.logs, move InputTypes to respective type files
